feat: 관광 문화 API 구현

+++ src/main/java/com/team/saver/market/store/dto/MarketResponse.java
@@ -36,6 +36,6 @@ public class MarketResponse {
 
     private long reviewCount;
 
-    private double maxDiscountRate;
+    private int maxDiscountRate;
 
 }
